# backend/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 3m-ui API
  version: "1.2"
  description: |
    REST API for the 3m-ui Mihomo management panel.

    - Base path: `/api/v1`
    - Admin routes need `Authorization: Bearer <jwt>` from `POST /auth/login`
    - Public subscription routes use path tokens (no JWT)

    Subscription URLs:
    - Always available: `/api/v1/client/sub/{token}` (and `/client/clash|v2ray|json/{token}`)
    - Optional short path from config `server.sub_path` (e.g. `/sub/{token}`) on the same
      process; optional `server.sub_port` serves subscription routes on a second listener.
    - Generated links use `server.public_url` + `SubscriptionBasePath` (sub_path or legacy).

    Cross-machine subscription: sync remote nodes via `/cluster/{id}/sync-nodes`,
    bind with `/users/{id}/remote-nodes`, then the user's sub URL merges local + remote.

    Proxy users support group/tags, per-user traffic reset cycle (`traffic_reset_days`)
    and post-expiry renew (`expire_renew_days`). Export all share links via
    `GET /users/export-links`.
servers:
  - url: /api/v1
...
paths:
...
  /cluster/{id}/mihomo/restart:
    post:
      tags: [cluster]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      summary: Restart Mihomo on remote
      responses:
        "200":
          description: OK / upstream error
...
  /mihomo/releases:
    get:
      tags: [mihomo]
      summary: List verifiable official stable and Pre Mihomo releases for this platform
      responses:
        "200":
          description: Release list with version, published_at, url, asset, sha256, size, and prerelease
        "502":
          description: Official release metadata is unavailable
  /mihomo/update:
    get:
      tags: [mihomo]
      summary: Get core update capability, saved previous version, and latest job
      responses:
        "200":
          description: supported, disabled_reason, busy, managed, previous_version, and job
    post:
      tags: [mihomo]
      summary: Start an asynchronous official core update (administrator only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: string
                  pattern: '^(v[0-9]+\.[0-9]+\.[0-9]+|alpha-[a-f0-9]{7,40})$'
      responses:
        "202":
          description: Job with id, operation, version, status, stage, and started_at; poll GET /mihomo/update
        "400":
          description: Invalid version or unavailable persistent storage/platform
        "409":
          description: Another core update or deployment is in progress
  /mihomo/update/rollback:
    post:
      tags: [mihomo]
      summary: Restore the saved previous core asynchronously (administrator only)
      responses:
        "202":
          description: Accepted rollback job; failures include error and rolled_back
        "400":
          description: No previous core is available or this platform is unsupported
        "409":
          description: Another update or deployment is in progress
...
components:
...
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: integer }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []